refactor: change counts to int64

+++ api/domain/protocols/Stats-Response.go
@@ -1,7 +1,7 @@
 package protocols
 
 type StatsResponse struct {
-	CountSimianDNA int `json:"count_simian_dna"`
-	CountHumanDNA int `json:"count_human_dna"`
+	CountSimianDNA int64 `json:"count_simian_dna"`
+	CountHumanDNA int64 `json:"count_human_dna"`
 	Ratio float64 `json:"ratio"`
 }
